Added custom problem checker
`server.py` listens to POST requests to a specified port, the competitive
companion browser extension posts parsed problems to this endpoint,
and creates a `samples` directory, for each testcase.

`run.py` compiles and runs the program against the parsed sample test
cases.

`run.py custom` compiles, and runs against user input.

`server.py` is used as a daemon process, in the background so it's
aliased as `cphd`.

`run.py` is used as a script, so it's aliased as `cpr`.